오류수정

+++ src/main/java/com/dksys/biz/user/ar/ar02/mapper/AR02Mapper.java
@@ -36,7 +36,7 @@ public interface AR02Mapper {
 
 	String selectOwner1ClntCd(Map<String, String> paramMap);
 
-	int selectBilgVatAmt(Map<String, String> paramMap);
+	long selectBilgVatAmt(Map<String, String> paramMap);
 
 	
 }
